+++ src/index.js
@@ -18,6 +18,9 @@ export { combination } from "./combinatorics/combination.js";
 
 // Function Functions
 export { carmichael } from "./function/carmichael.js";
+export { isAbundant } from "./function/isAbundant.js";
+export { isDeficient } from "./function/isDeficient.js";
+export { isPerfect } from "./function/isPerfect.js";
 export { mobius } from "./function/mobius.js";
 export { productOfDigits } from "./function/productOfDigits.js";
 export { sigma } from "./function/sigma.js";